Move #load from Console to Loader < Console; add loader_spec.rb
