+++ lib/console.rb
@@ -1,16 +1,24 @@
 class Console
   include BankErrors
+  include Console::CardPicker
   include Console::MoneyOperationsConsole
-  include Console::Creater
-  include Console::Loader
+  include Console::Creator
+  # include Console::Loader
 
-  # attr_accessor :current_account
+  MENU_COMMANDS = {
+    SC: :show_cards,
+    CC: :create_card,
+    DC: :destroy_card,
+    PM: :put_money,
+    WM: :withdraw_money,
+    SM: :send_money,
+    DA: :destroy_account
+  }.freeze
 
   def initialize
     @errors = []
     @current_account = nil
     account
-    # menu_commands
   end
 
   def account
@@ -32,36 +40,12 @@ class Console
   end
 
   def load
-    return create_the_first_account unless accounts.any?
-
-    puts I18n.t(:enter_login)
-    login = gets.chomp
-    puts I18n.t(:enter_password)
-    password = gets.chomp
-
-    account.load(login, password)
-    current_account
-    main_menu
+    Loader.new.load
   end
 
   def current_account
     @current_account ||= account.current_account
   end
-
-=begin
-  def create_the_first_account
-    puts I18n.t(:create_first_account)
-    gets.chomp == 'y' ? create : console
-  end
-
-  def create
-    ask_credentials
-
-    account.create
-    @current_account = account.current_account
-    main_menu
-  end
-=end
 
   def main_menu
     loop do
@@ -78,33 +62,10 @@ class Console
     retry
   end
 
-  MENU_COMMANDS = {
-    SC: :show_cards,
-    CC: :create_card,
-    DC: :destroy_card,
-    PM: :put_money,
-    WM: :withdraw_money,
-    SM: :send_money,
-    DA: :destroy_account
-  }.freeze
-
   def commands(command)
     raise CommandError if MENU_COMMANDS[command.to_sym].nil?
 
     method(MENU_COMMANDS[command.to_sym]).call
-=begin
-    case command
-    when 'SC' then show_cards
-    when 'CC' then create_card
-    when 'DC' then destroy_card
-    when 'PM' then put_money
-    when 'WM' then withdraw_money
-    when 'SM' then send_money
-    when 'DA' then destroy_account
-    # else puts I18n.t(:wrong_command)
-    else raise CommandError
-    end
-=end
   end
 
   def show_cards
@@ -150,95 +111,12 @@ class Console
 
   private
 
-  def ask_credentials
-    loop do
-      account.name = name_input
-      account.age = age_input
-      account.login = login_input
-      account.password = password_input
-      break if @errors.empty?
-
-      return_errors
-    end
-  end
-
-  def name_input
-    puts I18n.t(:enter_name)
-    name = gets.chomp
-    return name if name != '' && name.capitalize == name
-
-    @errors.push(I18n.t(:name_must_be))
-  end
-
-  def age_input
-    puts I18n.t(:enter_age)
-    age = gets.chomp
-    return age.to_i if age.to_i.is_a?(Integer) && age.to_i >= 23 && age.to_i <= 90
-
-    @errors.push(I18n.t(:age_must_be))
-  end
-
-  def login_input
-    puts I18n.t(:enter_login)
-    login = gets.chomp
-
-    @errors.push(I18n.t(:login_must_be)) if login == ''
-    @errors.push(I18n.t(:login_must_longer)) if login.length < 4
-    @errors.push(I18n.t(:login_must_shorter)) if login.length > 20
-
-    login_exists?(login) ? @errors.push(I18n.t(:account_exists)) : login
-  end
-
   def login_exists?(login)
     account.accounts.map(&:login).include? login
   end
 
-  #### MOVE errors into app?
-  def password_input
-    puts I18n.t(:enter_password)
-    password = gets.chomp
-
-    @errors.push(I18n.t(:password_must_be)) if password == ''
-    @errors.push(I18n.t(:password_must_longer)) if password.length < 6
-    @errors.push(I18n.t(:password_must_shorter)) if password.length > 30
-
-    password
-  end
-
-  def choose_the_card(operation)
-    # @current_account.card.any? ? (puts I18n.t(:choose_card, action: operation)) : (raise NoActiveCard)
-    puts I18n.t(:choose_card, action: operation)
-    show_cards_with_index
-
-    puts I18n.t(:press_exit)
-    answer_card = gets.chomp
-    return answer_card if answer_card == 'exit'
-
-    card_exists?(answer_card) ? answer_card : (raise WrongNumberError)
-  end
-
-  def card_exists?(answer_card)
-    (1..@current_account.card.length).include? answer_card.to_i
-  end
-
-  def show_cards_with_index
-    raise NoActiveCard if @current_account.card.empty?
-
-    @current_account.card.each_with_index do |card, index|
-      puts I18n.t(:show_card, num: card.number, type: card.type, index: index + 1)
-    end
-  end
-
   def accounts
     account.accounts
-  end
-
-  def current_card(number)
-    @current_account.card[number.to_i - 1]
-  end
-
-  def current_account_cards
-    @current_account.card
   end
 
   def return_errors
