+++ spec/console/loader_spec.rb
@@ -1,0 +1,74 @@
+RSpec.describe Loader do  
+  ASK_PHRASES = {
+    name: 'Enter your name',
+    login: 'Enter your login',
+    password: 'Enter your password',
+    age: 'Enter your age'
+  }.freeze
+
+  ERROR_PHRASES = {
+    user_not_exists: 'There is no account with given credentials',
+    wrong_command: 'Wrong command. Try again!',
+    no_active_cards: "There is no active cards!",
+    wrong_card_type: "Wrong card type. Try again!\n",
+    wrong_number: "You entered wrong number!\n",
+    correct_amount: 'You must input correct amount of money',
+    tax_higher: 'Your tax is higher than input amount'
+  }.freeze
+  
+  let(:current_subject) { described_class.new }
+
+  before do
+    current_subject.account.instance_variable_set(:@file_path, OVERRIDABLE_FILENAME)
+  end
+
+  describe '#load' do
+    context 'without active accounts' do
+      it do
+        expect(current_subject).to receive(:accounts).and_return([])
+        expect(current_subject).to receive(:create_the_first_account).and_return([])
+        current_subject.load
+      end
+    end
+
+    context 'with active accounts' do
+      let(:login) { 'Johnny' }
+      let(:password) { 'johnny1' }
+
+      before do
+        allow(current_subject).to receive_message_chain(:gets, :chomp).and_return(*all_inputs)
+        allow_any_instance_of(Account).to receive(:accounts) { [instance_double('Account', login: login, password: password)] }
+      end
+
+      context 'with correct outout' do
+        let(:all_inputs) { [login, password] }
+
+        it do
+          expect(current_subject).to receive(:main_menu)
+          [ASK_PHRASES[:login], ASK_PHRASES[:password]].each do |phrase|
+            expect(current_subject).to receive(:puts).with(phrase)
+          end
+          current_subject.load
+        end
+      end
+
+      context 'when account exists' do
+        let(:all_inputs) { [login, password] }
+
+        it do
+          expect(current_subject).to receive(:main_menu)
+          expect { current_subject.load }.not_to output(/#{ERROR_PHRASES[:user_not_exists]}/).to_stdout
+        end
+      end
+
+      context 'when account doesn\t exists' do
+        let(:all_inputs) { ['test', 'test', login, password] }
+
+        it do
+          expect { current_subject.load }.to raise_error(BankErrors::NoAccountError)
+          expect(current_subject).to receive(:console)
+        end
+      end
+    end
+  end
+end
