+++ spec/console_spec.rb
@@ -260,7 +260,7 @@ RSpec.describe Console do
       end
     end
   end
-
+=begin
   describe '#load' do
     context 'without active accounts' do
       it do
@@ -310,6 +310,7 @@ RSpec.describe Console do
       end
     end
   end
+=end
 
   describe '#create_the_first_account' do
     let(:cancel_input) { 'sdfsdfs' }
